페이징 쿼리 수정 및 css 수정 및 결제 상제페이지 데이터 불러오기 !

+++ src/com/kh/lp/money/model/vo/MoneyDetail.java
@@ -16,7 +16,7 @@ public class MoneyDetail {
 	private String brand;
 	private String model;
 	private int price;
-	private int priceFee;
+	private double priceFee;
 	
 
 }
